+++ test/nuts/local/localTrackingScenario.nut.ts
@@ -5,11 +5,11 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { EOL } from 'node:os';
-import * as path from 'node:path';
-import * as fs from 'node:fs';
+import path from 'node:path';
+import fs from 'node:fs';
 import { TestSession } from '@salesforce/cli-plugins-testkit';
 import { expect } from 'chai';
-import { shouldThrow } from '@salesforce/core/lib/testSetup';
+import { shouldThrow } from '@salesforce/core/testSetup';
 import { ShadowRepo } from '../../../src/shared/localShadowRepo';
 
 describe('end-to-end-test for local tracking', () => {